Move to Kotlin 2.0.0 and kotlin/multiplatform build.
Use RefCallback rather than useRefCallback for ref callbacks.
